[#3835] extracted mappers
