+++ GAE/src/org/akvo/flow/rest/form/FormAssemblyRestService.java
@@ -22,36 +22,20 @@ package org.akvo.flow.rest.form;
 import com.gallatinsystems.common.domain.UploadStatusContainer;
 import com.gallatinsystems.common.util.S3Util;
 import com.gallatinsystems.common.util.ZipUtil;
-import com.gallatinsystems.survey.dao.CascadeResourceDao;
 import com.gallatinsystems.survey.dao.SurveyDAO;
 import com.gallatinsystems.survey.dao.SurveyGroupDAO;
 import com.gallatinsystems.survey.dao.SurveyUtils;
-import com.gallatinsystems.survey.dao.TranslationDao;
-import com.gallatinsystems.survey.domain.CascadeResource;
 import com.gallatinsystems.survey.domain.Question;
-import static com.gallatinsystems.survey.domain.Question.Type.CASCADE;
-import com.gallatinsystems.survey.domain.QuestionGroup;
-import com.gallatinsystems.survey.domain.QuestionOption;
 import com.gallatinsystems.survey.domain.Survey;
 import com.gallatinsystems.survey.domain.SurveyGroup;
-import com.gallatinsystems.survey.domain.Translation;
 import com.gallatinsystems.survey.domain.WebForm;
-import com.google.appengine.api.datastore.KeyFactory;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Properties;
-import java.util.TreeMap;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -59,9 +43,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.server.ResponseStatusException;
-import org.waterforpeople.mapping.app.gwt.client.survey.QuestionDto;
-import org.waterforpeople.mapping.app.gwt.client.survey.QuestionGroupDto;
-import org.waterforpeople.mapping.app.gwt.client.survey.QuestionOptionDto;
 import org.waterforpeople.mapping.app.gwt.client.survey.SurveyDto;
 
 @Controller
@@ -71,15 +52,18 @@ public class FormAssemblyRestService {
     private static final Logger log = Logger.getLogger(FormAssemblyRestService.class.getName());
     private static final String SURVEY_UPLOAD_DIR = "surveyuploaddir";
     private static final String SURVEY_UPLOAD_URL = "surveyuploadurl";
-    public static final String BUCKET = "s3bucket";
+    private static final String BUCKET = "s3bucket";
+
     private final XmlFormAssembler xmlFormAssembler = new XmlFormAssembler();
+    private final FormDtoMapper formDtoMapper = new FormDtoMapper(new QuestionGroupDtoMapper(new QuestionDtoMapper(new QuestionOptionDtoMapper())),
+            new TranslationsDtoMapper());
 
     @PostMapping(consumes = "application/json")
     @ResponseBody
     public SurveyDto publishForm(@RequestBody SurveyDto surveyDto) {
         SurveyGroup survey = new SurveyGroupDAO().getByKey(surveyDto.getSurveyGroupId());
         long formId = surveyDto.getKeyId();
-        Survey form = assembleForm(surveyDto);
+        Survey form = formDtoMapper.assembleForm(surveyDto);
         try {
             FormUploadXml formUploadXml = xmlFormAssembler.assembleXmlForm(survey, form);
             log.info("Uploading " + formId);
@@ -103,16 +87,8 @@ public class FormAssemblyRestService {
         }
     }
 
-    private Survey assembleForm(SurveyDto surveyDto) {
-        Survey form = getFormFromDto(surveyDto);
-        attachFormTranslations(form);
-        List<Question> questions = getQuestionList(form.getQuestionGroupMap());
-        attachCascadeResources(questions);
-        return form;
-    }
-
     private void formUploadSuccess(SurveyGroup survey, Survey form) {
-        List<Question> questions = getQuestionList(form.getQuestionGroupMap());
+        List<Question> questions = formDtoMapper.getQuestionList(form.getQuestionGroupMap());
         form.setStatus(Survey.Status.PUBLISHED);
         if (form.getWebForm()) {
             boolean webForm = WebForm.validWebForm(survey, form, questions);
@@ -124,219 +100,6 @@ public class FormAssemblyRestService {
         List<Long> ids = new ArrayList<>();
         ids.add(form.getObjectId());
         SurveyUtils.notifyReportService(ids, "invalidate");
-    }
-
-    private Survey getFormFromDto(SurveyDto surveyDto) {
-        Survey form = new Survey();
-        form.setKey(KeyFactory.createKey("Survey", surveyDto.getKeyId()));
-        form.setCode(surveyDto.getCode());
-        form.setName(surveyDto.getName());
-        form.setVersion(Double.parseDouble(surveyDto.getVersion()));
-        form.setDesc(surveyDto.getDescription());
-        form.setStatus(Survey.Status.NOT_PUBLISHED);
-        form.setPath(surveyDto.getPath());
-        form.setSurveyGroupId(surveyDto.getSurveyGroupId());
-        form.setDefaultLanguageCode(surveyDto.getDefaultLanguageCode());
-        form.setRequireApproval(surveyDto.getRequireApproval());
-        form.setCreatedDateTime(surveyDto.getCreatedDateTime());
-        form.setLastUpdateDateTime(surveyDto.getLastUpdateDateTime());
-        form.setAncestorIds(surveyDto.getAncestorIds());
-        TreeMap<Integer, QuestionGroup> groupMap = getGroupMap(surveyDto);
-        form.setQuestionGroupMap(groupMap);
-        return form;
-    }
-
-    @Nonnull
-    private TreeMap<Integer, QuestionGroup> getGroupMap(SurveyDto surveyDto) {
-        TreeMap<Integer, QuestionGroup> groupMap = new TreeMap<>();
-        List<QuestionGroupDto> groupDtos = surveyDto.getQuestionGroupList();
-        if (groupDtos != null) {
-            int i = 1;
-            for (QuestionGroupDto groupDto : groupDtos) {
-                QuestionGroup group = mapToGroup(groupDto);
-                if (groupMap.containsKey(groupDto.getOrder())) {
-                    groupMap.put(i, group);
-                    groupDto.setOrder(i);
-                } else {
-                    int order = groupDto.getOrder() != null ? groupDto.getOrder() : i;
-                    groupMap.put(order, group);
-                }
-                i++;
-            }
-        }
-        return groupMap;
-    }
-
-    private QuestionGroup mapToGroup(QuestionGroupDto groupDto) {
-        QuestionGroup group = new QuestionGroup();
-        group.setKey(KeyFactory.createKey("QuestionGroup", groupDto.getKeyId()));
-        group.setCode(groupDto.getCode());
-        group.setSurveyId(groupDto.getSurveyId());
-        group.setOrder(groupDto.getOrder());
-        group.setPath(groupDto.getPath());
-        group.setName(groupDto.getName());
-        group.setRepeatable(groupDto.getRepeatable());
-        group.setStatus(QuestionGroup.Status.valueOf(groupDto.getStatus()));
-        group.setImmutable(groupDto.getImmutable());
-        group.setQuestionMap(mapQuestions(groupDto));
-        return group;
-    }
-
-    @Nonnull
-    private TreeMap<Integer, Question> mapQuestions(QuestionGroupDto groupDto) {
-        TreeMap<Integer, Question> mappedQuestions = new TreeMap<>();
-        List<QuestionDto> questionList = groupDto.getQuestionList();
-        if (questionList != null) {
-            for (QuestionDto questionDto : questionList) {
-                mappedQuestions.put(questionDto.getOrder(), mapToQuestion(questionDto));
-            }
-        }
-        return mappedQuestions;
-    }
-
-    private Question mapToQuestion(QuestionDto questionDto) {
-        Question question = new Question();
-        question.setKey(KeyFactory.createKey("Question", questionDto.getKeyId()));
-        question.setType(Question.Type.valueOf(questionDto.getType().toString()));
-        question.setTip(questionDto.getTip());
-        question.setText(questionDto.getText());
-        question.setDependentFlag(questionDto.getDependentFlag());
-        question.setAllowMultipleFlag(questionDto.getAllowMultipleFlag());
-        question.setAllowOtherFlag(questionDto.getAllowOtherFlag());
-        question.setCollapseable(questionDto.getCollapseable());
-        question.setGeoLocked(questionDto.getGeoLocked());
-        question.setRequireDoubleEntry(questionDto.getRequireDoubleEntry());
-        question.setImmutable(questionDto.getImmutable());
-        question.setDependentQuestionId(questionDto.getDependentQuestionId());
-        question.setDependentQuestionAnswer(questionDto.getDependentQuestionAnswer());
-        question.setCascadeResourceId(questionDto.getCascadeResourceId());
-        question.setCaddisflyResourceUuid(questionDto.getCaddisflyResourceUuid());
-        question.setQuestionGroupId(question.getQuestionGroupId());
-        question.setSurveyId(questionDto.getSurveyId());
-        question.setVariableName(questionDto.getVariableName());
-        question.setOrder(questionDto.getOrder());
-        question.setMandatoryFlag(questionDto.getMandatoryFlag());
-        question.setPath(questionDto.getPath());
-        question.setAllowDecimal(questionDto.getAllowDecimal());
-        question.setAllowSign(questionDto.getAllowSign());
-        question.setMinVal(questionDto.getMinVal());
-        question.setMaxVal(question.getMaxVal());
-        question.setAllowExternalSources(questionDto.getAllowExternalSources());
-        question.setLocaleNameFlag(questionDto.getLocaleNameFlag());
-        question.setLocaleLocationFlag(questionDto.getLocaleLocationFlag());
-        question.setPersonalData(questionDto.getPersonalData());
-        question.setAllowPoints(questionDto.getAllowPoints());
-        question.setAllowLine(questionDto.getAllowLine());
-        question.setAllowPolygon(questionDto.getAllowPolygon());
-        question.setSourceQuestionId(questionDto.getSourceId());
-        question.setQuestionOptionMap(mapToOptions(questionDto));
-        return question;
-    }
-
-    @Nonnull
-    private TreeMap<Integer, QuestionOption> mapToOptions(QuestionDto questionDto) {
-        TreeMap<Integer, QuestionOption> mappedOptions = new TreeMap<>();
-        List<QuestionOptionDto> dtoList = questionDto.getOptionList();
-        if (dtoList != null) {
-            for (QuestionOptionDto questionOptionDto : dtoList) {
-                mappedOptions.put(questionOptionDto.getOrder(), mapToQuestionOption(questionOptionDto));
-            }
-        }
-        return mappedOptions;
-    }
-
-    private QuestionOption mapToQuestionOption(QuestionOptionDto questionOptionDto) {
-        QuestionOption questionOption = new QuestionOption();
-        questionOption.setKey(KeyFactory.createKey("QuestionOption", questionOptionDto.getKeyId()));
-        questionOption.setText(questionOptionDto.getText());
-        questionOption.setCode(questionOptionDto.getCode());
-        questionOption.setOrder(questionOptionDto.getOrder());
-        questionOption.setQuestionId(questionOptionDto.getQuestionId());
-        return questionOption;
-    }
-
-    private void attachCascadeResources(List<Question> questions) {
-        CascadeResourceDao cascadeResourceDao = new CascadeResourceDao();
-        for (Question question : questions) {
-            if (CASCADE.equals(question.getType())) {
-                CascadeResource cascadeResource = cascadeResourceDao.getByKey(question.getCascadeResourceId());
-                if (cascadeResource != null) {
-                    question.setCascadeResource(cascadeResource.getResourceId());
-                    question.setLevelNames(cascadeResource.getLevelNames());
-                }
-            }
-        }
-    }
-
-    private void attachFormTranslations(Survey form) {
-        Map<Long, List<Translation>> translations = new TranslationDao().mappedTranslationsByParentId(form.getObjectId());
-        HashMap<String, Translation> formTranslationMap = getMappedTranslationsForParent(translations, form.getObjectId());
-        if (formTranslationMap.size() > 0) {
-            form.setTranslationMap(formTranslationMap);
-        }
-        TreeMap<Integer, QuestionGroup> questionGroupMap = form.getQuestionGroupMap();
-        if (questionGroupMap != null) {
-            List<QuestionGroup> groups = new ArrayList<>(questionGroupMap.values());
-            for (QuestionGroup group : groups) {
-                HashMap<String, Translation> map = getMappedTranslationsForParent(translations, group.getKey().getId());
-                if (map.size() > 0) {
-                    group.setTranslations(map);
-                }
-                TreeMap<Integer, Question> questionMap = group.getQuestionMap();
-                if (questionMap != null) {
-                    List<Question> questions = new ArrayList<>(questionMap.values());
-                    for (Question question : questions) {
-                        List<Translation> questionTranslations = translations.get(question.getKey().getId());
-                        if (questionTranslations.size() > 0) {
-                            question.setTranslations(questionTranslations);
-                        }
-                        TreeMap<Integer, QuestionOption> questionOptionMap = question.getQuestionOptionMap();
-                        if (questionOptionMap != null) {
-                            List<QuestionOption> options = new ArrayList<>(questionOptionMap.values());
-                            for (QuestionOption option : options) {
-                                HashMap<String, Translation> optionTranslationsMap = getMappedTranslationsForParent(translations, option.getKey().getId());
-                                if (optionTranslationsMap.size() > 0) {
-                                    option.setTranslationMap(optionTranslationsMap);
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    private HashMap<String, Translation> getMappedTranslationsForParent(Map<Long, List<Translation>> translations, Long parentId) {
-        List<Translation> translationsForForm = translations.get(parentId);
-        return mapTranslations(translationsForForm);
-    }
-
-    private HashMap<String, Translation> mapTranslations(@Nullable List<Translation> translations) {
-        HashMap<String, Translation> mappedTranslations = new HashMap<>();
-        if (translations != null) {
-            for (Translation t : translations) {
-                mappedTranslations.put(t.getLanguageCode(), t);
-            }
-        }
-        return mappedTranslations;
-    }
-
-    @Nonnull
-    private List<Question> getQuestionList(@Nullable TreeMap<Integer, QuestionGroup> questionGroupTreeMap) {
-        if (questionGroupTreeMap != null) {
-            List<QuestionGroup> groups = new ArrayList<>(questionGroupTreeMap.values());
-            return groups.stream().filter(it -> it.getQuestionMap() != null).map(this::getValues).collect(Collectors.toList())
-                    .stream().flatMap(Collection::stream).collect(Collectors.toList());
-        }
-        return Collections.emptyList();
-    }
-
-    private List<Question> getValues(QuestionGroup questionGroup) {
-        TreeMap<Integer, Question> questionMap = questionGroup.getQuestionMap();
-        if (questionMap == null) {
-            return Collections.emptyList();
-        }
-        return new ArrayList<>(questionMap.values());
     }
 
     /**
@@ -369,5 +132,4 @@ public class FormAssemblyRestService {
             return false;
         }
     }
-
 }
